Move resource response sender utils to resource package
