+++ backend/resource/doc.go
@@ -1,0 +1,2 @@
+// Package resource provides utils for handling resource calls.
+package resource
